+++ azext_iot/tests/iothub/configurations/test_iot_config_int.py
@@ -7,7 +7,7 @@
 import random
 import json
 
-from azext_iot.tests import IoTLiveScenarioTest
+from azext_iot.tests.iothub import IoTLiveScenarioTest
 from azext_iot.tests.conftest import get_context_path
 from azext_iot.tests.iothub import DATAPLANE_AUTH_TYPES
 from azext_iot.common.utility import read_file_content
