+++ azext_iot/tests/iothub/modules/test_iothub_modules_int.py
@@ -4,7 +4,7 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-from azext_iot.tests import IoTLiveScenarioTest
+from azext_iot.tests.iothub import IoTLiveScenarioTest
 from time import sleep
 from azext_iot.common.utility import generate_key
 from azext_iot.tests.iothub import (
